% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class.R
\name{get_classes}
\alias{get_classes}
\title{Get class function}
\usage{
get_classes(x)
}
\arguments{
\item{R}{object from which you want to determin the classes, can be a list of objects or a data frame}
}
\value{
dataframe - in case a list is provided as input for every element in the list the function returns its class,
in case dataframe is an input the class is return for every column.
}
\description{
Get class function
}

