% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_list.R
\name{json_to_list}
\alias{json_to_list}
\title{json_to_list - function to convert a columncontaining json to a list in R}
\usage{
json_to_list(data, column_name)
}
\arguments{
\item{data}{- dataframe}

\item{column_name}{- column name containg json}
}
\value{
list
}
\description{
json_to_list - function to convert a columncontaining json to a list in R
}

